Bump grpc-java to 1.18.0
Requires bumping other dependencies in deps.json such as Netty (4.1.32)
and error_prone_annotations. Includes also fixes detected by updated
errorprone.

+++ apps/virtual/app/src/main/java/org/onosproject/incubator/net/virtual/store/impl/SimpleVirtualFlowRuleStore.java
@@ -60,7 +60,7 @@ import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 import static org.onosproject.incubator.net.virtual.store.impl.OsgiPropertyConstants.PENDING_FUTURE_TIMEOUT_MINUTES;
 import static org.onosproject.incubator.net.virtual.store.impl.OsgiPropertyConstants.PENDING_FUTURE_TIMEOUT_MINUTES_DEFAULT;
@@ -91,12 +91,12 @@ public class SimpleVirtualFlowRuleStore
             flowEntries = new ConcurrentHashMap<>();
 
 
-    private final AtomicInteger localBatchIdGen = new AtomicInteger();
+    private final AtomicLong localBatchIdGen = new AtomicLong();
 
     /** Expiration time after an entry is created that it should be automatically removed. */
     private int pendingFutureTimeoutMinutes = PENDING_FUTURE_TIMEOUT_MINUTES_DEFAULT;
 
-    private Cache<Integer, SettableFuture<CompletedBatchOperation>> pendingFutures =
+    private Cache<Long, SettableFuture<CompletedBatchOperation>> pendingFutures =
             CacheBuilder.newBuilder()
                     .expireAfterWrite(pendingFutureTimeoutMinutes, TimeUnit.MINUTES)
                     .removalListener(new TimeoutFuture())
@@ -119,7 +119,7 @@ public class SimpleVirtualFlowRuleStore
         readComponentConfiguration(context);
 
         // Reset Cache and copy all.
-        Cache<Integer, SettableFuture<CompletedBatchOperation>> prevFutures = pendingFutures;
+        Cache<Long, SettableFuture<CompletedBatchOperation>> prevFutures = pendingFutures;
         pendingFutures = CacheBuilder.newBuilder()
                 .expireAfterWrite(pendingFutureTimeoutMinutes, TimeUnit.MINUTES)
                 .removalListener(new TimeoutFuture())
@@ -213,7 +213,7 @@ public class SimpleVirtualFlowRuleStore
         }
 
         SettableFuture<CompletedBatchOperation> r = SettableFuture.create();
-        final int futureId = localBatchIdGen.incrementAndGet();
+        final long futureId = localBatchIdGen.incrementAndGet();
 
         pendingFutures.put(futureId, r);
 
@@ -396,9 +396,9 @@ public class SimpleVirtualFlowRuleStore
     }
 
     private static final class TimeoutFuture
-            implements RemovalListener<Integer, SettableFuture<CompletedBatchOperation>> {
+            implements RemovalListener<Long, SettableFuture<CompletedBatchOperation>> {
         @Override
-        public void onRemoval(RemovalNotification<Integer,
+        public void onRemoval(RemovalNotification<Long,
                 SettableFuture<CompletedBatchOperation>> notification) {
             // wrapping in ExecutionException to support Future.get
             if (notification.wasEvicted()) {
